# server/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /product:
    # binds a127 app logic to a route
    x-swagger-router-controller: product
    post:
      description: Post subscription for the product for an user
      # used as the method name of the controller
      operationId: subscribe
      consumes:
        - application/json
      parameters:
        - in: body
          name: product
          description: The user to validate.
          schema:
            $ref: '#/definitions/Product'
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ProductResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /subscription:
    # binds a127 app logic to a route
    x-swagger-router-controller: product
    post:
        description: Get subscription for the product for an user
        # used as the method name of the controller
        operationId: getSubscription
        consumes:
          - application/json
        parameters:
          - in: body
            name: product
            description: To get the subscriptions
            schema:
              $ref: '#/definitions/Subscription'
        responses:
          "200":
            description: Success
            schema:
              # a pointer to a definition
              $ref: "#/definitions/HelloWorldResponse"
          # responses may fall through to errors
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  Product:
    type: object
    required:
      - productID
      - userID
    properties:
      productID:
        type: array
        items:
          type: string
      userID:
        type: string
  Subscription:
    type: object
    required:
      - userID
    properties:
      userID:
        type: string
...
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  ProductResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string